Viewer Updates
Viewer wasn't really a role before, added some updates that made it a
role. When viewer joins, they are given the current "READY" status of
the two player, and no buttons to click. The "READY" statuses update as
players 1&2 change them, and when the game is being played, the viewer
is able to see all cards drawn. This is obviously an issue with more
serious players, since you could just add a new tab to see your
opponent's cards. Not an issue currently between friends.