Client-Side Refactoring
Refactored the client side code, it was way too specific, and had a lot
of variables that were sitting around doing nothing. Generalized it so
that each client has one object for its ready/draw/reroll button that is
initialized when that player connects. Also added a reroll button
(currently does nothing), and started making some general functions that
will be used by/for cards.