Added Card/Ban Lists
Player 1 can now choose the ban/card list to go off of for the
randomization. Reworked the backend a little bit to be more generic, and
added code to have it keep track of each players "allowed" cards left
(Pot of Greed is limited, so 1 is allowed, each player is given a
starting value of 1 for Pot of Greed. If the card is added, that value
goes down by 1 to 0, and the program will know to not give Pot of Greed
to that player again. Rerolling will increment the value by 1). Did some
refactoring on the client side, to keep the code more concise and
readable.