CardType Specifying
Player 1 is now able to specify how many monster/trap/spell cards each
player's final deck will have. Updated the randomCard function to just
call the randomCardWeighted function with a randomFactor of 1. Added
cardData, which contains cardTypes (spell, monster, trap) and more
specific type (XYZ, Fusion, Equip, etc) (for future use). It also
contains whether the card is used in the extra deck or not, so that
those aren't included in the main deck. Reworked the backend quite a
bit, now there is an object for both players that contains data about
them, like rerolls and things that abilities may change in the future.
Removed the card list from client end, since they have their whole list
in the card div if they really need to get to it.