Abilities Added
Some basic abilities have been added, just three for now. Ability object
seems generic enough that adding more in the future should not be too
difficult. Need to update the ability server function so that it returns
an object with the ids of the abilities that are/aren't able to be used
now, so that I can actually show when an ability is available for use.
Currently there's no indication of that, or any indication that an
ability was used. Will need to do some commenting/refactoring after that