Potential First Working Version
Added some extra logic to the abilities, so that the person whose turn
it isn't won't be able to click, and players can now see when abilities
are not available. Game knows to end when card total is hit, and then
players can click a button to download the ydk file. Should be testing
with friends tomorrow, most of the next coding sessions will be
refactoring to help in future updates.