Bug Fixing & Refactoring
Ran some tests with the boys tonight, and caught a few bugs. Most
prominent was that if an ability was used to make you draw your
opponents last card, and for whatever reason you weren't able to draw
that card, drawCard was never set so a null card was given. Corrected
that, and found some issues with previously global variables
(randomFactor, currentPlayer, etc) still being referenced. Changed them
to be stored in roomData. Removed the cardsLeft object from the
playerData in the interest of space, just running a normal count now to
see how many more of a card could be added. Seems to be running fine.
Did a lot of refactoring/cleaning up/switching from let to const.