getRandomCard Bug Fix and Other Fixes
getRandomCard function, turns out, was the one throwing the errors when
we were playing (thanks, forever logs!). Looks like the error happened
when weights was never added to, meaning that lastId would be undefined,
and trying to get the cardType of undefined is not liked very much by
javscript, fun fact. Hopefully that and the null issue will solve all
future issues? Function will now log information if lastId ends up
having to be used, to hopefully understand why weights would ever be 0,
since it shouldn't be. Maybe the rework of the cardCount function
already solved it.