Bug Fixing and Log Addition
Looks like I tracked down all of the issues that were happening causing
crashes. The cardType counts were not always being updated (rerolls,
getting your/opponents card rather than a random one), which meant that
eventually the player just ran out of viable cards because no type was
allowed. Added a function that counts how many of each type is in the
deck for both players, and changed it so there are typeUsed and
typeAllowed properties in roomData now to keep track. Now the lastId
fallback of the randomCard functions should never have to be used. Also
added some basic logging, just to get basic events that are happening,
but it is just scratching the surface, and more will have to be added
for future use. Currently just want to this pushed and uploaded to the
server to let the boys test it.