Bug Fixing and Adding Russian Roulette
Breaking reroll and draw out into functions caused some issues, but
fixed them with this update. The reroll function also allowed a new
ability, Russian Roulette, where upon activation a random card of yours
will be rerolled. To do this, added some extra lines when updated stats
based on abilities that let it call a function, and putting the function
in the ability was super straight forward.