More Ability Functionality
Added more structure to abilities. They are now able to run custom
functions, which when passed the roomName and player should be able to
do just about anything (will need to think about how I want to handle an
ability that allows some kind of choice/should proc on opponents turn).
Abilities are now able to have a delay put on them, opening up even more
options for what they do. Also fixed an issue where abilities would
technically be able to be used when they weren't supposed to, since the
check at the beginning of the ability step only checked if the player was
the current one going, and the player didn't have a delay on their
ability usage. Now the function that updated the ability statuses
(blurring out non-usable abilities) also sets that ability's canUse
property, so it can be checked when an ability is pressed.