Background Color Removal/Deck Stats
Removed all of the ugly colors I had originally used to verify the
correct div locations, so now the generic background is just shown on
every screen. Also added a function/socket emit for the client & server
to share information on the card type makeup of the current deck, to
display to the player.