Playing Screen Change & Minor Updates
Did a total overhaul of the playing screen look. Got rid of the two
sided nature, and made it so it is just showing your side. Will also be
adding more information such as a log of events, better card type count
information, and card description of hovered card. Download screen is
gone, and download button now just appears on the same screen when all
cards are drawn. Final card can be rerolled with your remaining rerolls,
no abilities can be used.
Also made it so that you can just click a game to join in, and added
default of 10 everything for card counts, to make testing faster.
Viewing mode is now totally broken, will need to make the necessary
changes to allow viewers again.