Reworked Loading Screen
Fully reworked the loading screen, so that the player1 side already has
all of the elements baked into the HTML, contained within 1 div. That
way, it is easy to hide the elements for player2, but more importantly,
should now be able to set all of those components immediately, and not
have to worry about them again if a player is kicked from a game.