Added More Steps to firstTimeSetup
Previously, there were some addEventListener steps that were run on
first start, that were just plopped right in the middle of a bunch of
functions. Pulled them from there, and added them to firstTimeSetup so
that it's easier to see all that's going on when a client first visits
the page. Small documentation changes/additions were also made to some
functions.