index.html Full Function Documentation
All functions within index.html now have documentation on what they do,
what parameters they take, what they return, and when events are emitted
(in the case of socket functions). Now it's just a matter of keeping up
with the documentation from now on, because that was a long, drawn out
process I don't want to do again.