Usage Delay Properties Added to Abilities
Closes #6
useAgainDelay and turnsUntilUseAgain have been added as properties to
abilities. useAgainDelay is the delay that will be added to an ability
after it is used, which is stored in turnsUntilUseAgain.
turnsUntilUseAgain can be set to a number > 0 at the start to prevent an
ability from being used right away, and if that is done and
useAgainDelay isn't set, it will only implement a delay in the
beginning, not after usage.