Div For Viewing Cards While Playing
Closes #13
In an effort to add more future functionality to the program, I created
a div that will be used to show a player cards. The cards shown would be
for abilities (ability that allows player to have a selection of
cards to choose from for them/their opponent, ability that lets them
view a random opponent card, etc.).