Updated Client/Server for Online Use
Because of how different the environments are between running this code
on the website vs my local machine, had to add some code to allow it to
run on an https server, as well as going through the correct website
paths for it to work online.

Code is made to be easily commented in/out to switch between the two
depending on where I'm running it. Code is set up for local running, can
be edited for website use.