Adding Comments, and Cleaning makeRoom
Added some comments to some of the global variables, and cleaned up some
of the code in makeRoom. A bunch of actions were happening within a for
loop that had no reason to be execued in the loop, and an extra variable
(username), had been created and used when it was completely extraneous.