Specified Reroll Emit, Cleaned setCardMouseEvents
Reroll emit was originally going to every person in the room, when only
the player that initiated it needs to be getting information from that
emit. The reroll socket.on was checking for the player in the beginning,
and immediately returning if it was viewer or not equal to myRole.

The function that was setting the card mouse events (showing the hovered
card in large format and displaying the limit on the card) originally
had to have limit passed in as a separate argument, but there's no real
need to do that. Instead, attached the limit to card at card.limit, and
that way the card has access to the information to it, so no more need
for passing the limit.